requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-environ==0.4.5
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djongo==1.3.4
dnspython==2.1.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pymongo==3.11.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
